refactor: loading UX for downloading file from URL
